pyproject.toml
[tool.poetry]
name = "home-controller-system"
version = "0.1.0"
description = ""
authors = ["Jonathen Sundy <jonathensundy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
poetrify = "^0.4.1"
python-socketio = "^4.6.0"
eventlet = "^0.25.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
